Move render test support to document-capture
**Why**: It's not generic render helper, since it's quite specific to the document capture package
