+++ app/javascript/packages/document-capture/components/acuant-capture.jsx
@@ -203,7 +203,7 @@ function AcuantCapture(
         bannerText={bannerText}
         invalidTypeText={t('errors.doc_auth.invalid_file_input_type')}
         fileUpdatedText={t('doc_auth.info.image_updated')}
-        accept={isMockClient ? undefined : ['image/jpeg', 'image/png', 'image/bmp', 'image/tiff']}
+        accept={['image/jpeg', 'image/png', 'image/bmp', 'image/tiff']}
         capture={capture}
         value={value}
         errorMessage={ownErrorMessage ?? errorMessage}
