+++ spec/javascripts/packages/document-capture/components/documents-step-spec.jsx
@@ -3,7 +3,7 @@ import userEvent from '@testing-library/user-event';
 import sinon from 'sinon';
 import DeviceContext from '@18f/identity-document-capture/context/device';
 import DocumentsStep from '@18f/identity-document-capture/components/documents-step';
-import render from '../../../support/render';
+import { render } from '../../../support/document-capture';
 
 describe('document-capture/components/documents-step', () => {
   it('renders with front and back inputs', () => {
