+++ spec/javascripts/packages/document-capture/components/mobile-intro-step-spec.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { I18nContext, ServiceProviderContext } from '@18f/identity-document-capture';
 import MobileIntroStep from '@18f/identity-document-capture/components/mobile-intro-step';
-import render from '../../../support/render';
+import { render } from '../../../support/document-capture';
 
 describe('document-capture/components/mobile-intro-step', () => {
   context('service provider context', () => {
