+++ spec/javascripts/packages/document-capture/components/submission-complete-spec.jsx
@@ -6,7 +6,7 @@ import SubmissionComplete, {
   RetrySubmissionError,
 } from '@18f/identity-document-capture/components/submission-complete';
 import SuspenseErrorBoundary from '@18f/identity-document-capture/components/suspense-error-boundary';
-import render from '../../../support/render';
+import { render } from '../../../support/document-capture';
 import { useCleanDOM } from '../../../support/dom';
 import { useSandbox } from '../../../support/sinon';
